remove Utilities/Discord folder
