+++ huisbot/Modules/Link.cs
@@ -2,7 +2,6 @@
 using huisbot.Models.Osu;
 using huisbot.Services;
 using huisbot.Utilities;
-using huisbot.Utilities.Discord;
 
 namespace huisbot.Modules;
